# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"UserWithRelationship": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"image": {
						"type": "string",
						"nullable": true
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"staff",
							"hoster"
						]
					},
					"banned": {
						"type": "boolean",
						"nullable": true
					},
					"customerCode": {
						"type": "string",
						"nullable": true
					},
					"lastLogin": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"userSettings": {
						"$ref": "#/components/schemas/UserSettings"
					},
					"organiser": {
						"$ref": "#/components/schemas/Organiser"
					},
					"accounts": {
						"items": {
							"$ref": "#/components/schemas/AccountOutput"
						},
						"type": "array"
					},
					"organizerPayoutAccount": {
						"$ref": "#/components/schemas/OrganizerPayoutAccount"
					}
				},
				"required": [
					"id",
					"name",
					"email",
					"emailVerified",
					"image",
					"role",
					"banned",
					"customerCode",
					"lastLogin",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_UserWithRelationship-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserWithRelationship"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"APIResponse_GetSSOAccounts-Array-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/GetSSOAccounts"
						},
						"type": "array",
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"APIResponse_User-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/User"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateUserInputBody": {
				"properties": {
					"image": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"user",
							"admin",
							"staff",
							"hoster"
						]
					},
					"banned": {
						"type": "boolean"
					},
					"emailVerified": {
						"type": "boolean"
					},
					"lastLogin": {
						"type": "string",
						"format": "date-time"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_UserSettings-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserSettings"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UpdateUserSettingsInputBody": {
				"properties": {
					"userId": {
						"type": "string"
					},
					"notification": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/NotificationSettings"
							},
							{}
						]
					},
					"preferences": {
						"type": "string",
						"enum": [
							"system",
							"dark",
							"light"
						]
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"BecomeOrganiserInputBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"organisationName": {
						"type": "string"
					},
					"phoneNumber": {
						"type": "string"
					},
					"phoneNumberVerified": {
						"type": "boolean"
					},
					"address": {
						"properties": {
							"country": {
								"type": "string"
							},
							"postalCode": {
								"type": "string"
							},
							"state": {
								"type": "string"
							},
							"city": {
								"type": "string"
							},
							"street": {
								"type": "string"
							}
						},
						"type": "object"
					},
					"logo": {
						"type": "string"
					},
					"bio": {
						"type": "string"
					},
					"website": {
						"type": "string"
					},
					"role": {
						"type": "string",
						"enum": [
							"admin",
							"staff",
							"hoster",
							"user"
						]
					}
				},
				"required": [
					"organisationName"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header",
				"bearerFormat": "JWT",
				"description": "Enter your bearer token in the format \"Bearer {token}\""
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "luminevent",
		"version": "1.2.1",
		"description": "Luminevent is a backend service for managing and organizing events, providing APIs for event creation, user management, and scheduling.",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "BITS"
		}
	},
	"paths": {
		"/users/me": {
			"get": {
				"operationId": "GetMe",
				"responses": {
					"200": {
						"description": "User data fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_UserWithRelationship-or-null_"
								}
							}
						}
					}
				},
				"summary": "Retrieves the details of the currently authenticated user.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": []
			}
		},
		"/users/{userId}": {
			"get": {
				"operationId": "GetUserById",
				"responses": {
					"200": {
						"description": "User fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_UserWithRelationship-or-null_"
								}
							}
						}
					}
				},
				"summary": "Retrieves a user by their ID.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateUser",
				"responses": {
					"200": {
						"description": "Updated profile data successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_User-or-null_"
								}
							}
						}
					}
				},
				"summary": "Updates a user's profile information.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserInputBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteUser",
				"responses": {
					"200": {
						"description": "User deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_null_"
								}
							}
						}
					}
				},
				"summary": "Deletes a user account.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}/social-signup-option-accounts": {
			"get": {
				"operationId": "GetUserSSOAccounts",
				"responses": {
					"200": {
						"description": "Social accounts fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_GetSSOAccounts-Array-or-null_"
								}
							}
						}
					}
				},
				"summary": "Retrieves all social accounts linked to a user.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}/settings": {
			"put": {
				"operationId": "UpdateUserSettings",
				"responses": {
					"200": {
						"description": "User settings updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_UserSettings-or-null_"
								}
							}
						}
					}
				},
				"summary": "Updates a user's settings.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateUserSettingsInputBody"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetUserSettings",
				"responses": {
					"200": {
						"description": "User settings fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_UserSettings-or-null_"
								}
							}
						}
					}
				},
				"summary": "Retrieves a user's settings.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/users/{userId}/become-organiser": {
			"post": {
				"operationId": "BecomeOrganiser",
				"responses": {
					"200": {
						"description": "User successfully became an organiser",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_UserWithRelationship-or-null_"
								}
							}
						}
					}
				},
				"summary": "Transforms an existing user into an organiser.",
				"tags": [
					"Users"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "userId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/BecomeOrganiserInputBody"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}